+++ tests/compilers/test_ast.py
@@ -16,7 +16,6 @@ from hy._compat import PY3
 import ast
 import pytest
 
-
 def _ast_spotcheck(arg, root, secondary):
     if "." in arg:
         local, full = arg.split(".", 1)
@@ -71,6 +70,17 @@ def test_empty_expr():
     "Empty expressions should be illegal at the top level."
     cant_compile("(print ())")
     can_compile("(print '())")
+
+
+def test_dot_unpacking():
+
+    can_compile("(.meth obj #* args az)")
+    cant_compile("(.meth #* args az)")
+    cant_compile("(. foo #* bar baz)")
+
+    can_compile("(.meth obj #** args az)")
+    can_compile("(.meth #** args obj)")
+    cant_compile("(. foo #** bar baz)")
 
 
 def test_ast_bad_if():
@@ -290,7 +300,7 @@ def test_ast_require():
 
 
 def test_ast_import_require_dotted():
-    """As in Python, it should be a compile-type error to attempt to
+    """As in Python, it should be a compile-time error to attempt to
 import a dotted name."""
     cant_compile("(import [spam [foo.bar]])")
     cant_compile("(require [spam [foo.bar]])")